--- ImplementationSuggestions.Fitofarmaceuticos/Readme.md
@@ -1,7 +1,0 @@
-# CustomTab Code Samples
-
-In this directory you will the code sample to implement the Implementation Suggetion that handles the "Fitofarmacêuticos" portuguese legal requirements in a PRIMAVERA ERP V10 environment.
-
-### How to use?
-Please refer to the Implementation Suggestion article avilable at: https://helpcenter.primaverabss.com/v10/como-configurar/extensibilidade/sugestoes-de-implementacao/ 
-
