+++ SIFitofarmaceuticos/ImplementationSuggestions.Fitofarmaceuticos/Properties/AssemblyInfo.cs
