+++ SIFitofarmaceuticos/ImplementationSuggestions.Fitofarmaceuticos/SIFitofarmaceuticos.csproj
@@ -8,8 +8,8 @@
     <ProjectGuid>{D7948B0F-1253-4A18-ABDD-48F59F724065}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>ImplementationSuggestions.Fitofarmaceuticos</RootNamespace>
-    <AssemblyName>ImplementationSuggestions.Fitofarmaceuticos</AssemblyName>
+    <RootNamespace>SIFitofarmaceuticos</RootNamespace>
+    <AssemblyName>Primavera.SIFitofarmaceuticos</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
   </PropertyGroup>
