+++ SIFitofarmaceuticos/Primavera.SIFitofarmaceuticos.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30611.23
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ImplementationSuggestions.Fitofarmaceuticos", "ImplementationSuggestions.Fitofarmaceuticos\ImplementationSuggestions.Fitofarmaceuticos.csproj", "{D7948B0F-1253-4A18-ABDD-48F59F724065}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SIFitofarmaceuticos", "ImplementationSuggestions.Fitofarmaceuticos\SIFitofarmaceuticos.csproj", "{D7948B0F-1253-4A18-ABDD-48F59F724065}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
